Added GetTransactionByMonth Command. Changed from double to int for year.

+++ Commands/GetTransactionsByYearCommand.cs
@@ -29,7 +29,7 @@ namespace EgenInlämning.Commands
            
             
 
-             double year = Convert.ToDouble(args[1]);
+             int year = Convert.ToInt32(args[1]);
              // Add function for choosing type of transaction to sum or show for specifc year.
             // if (args.Length > 0 || !int.TryParse(args[0], out year))
             // {
